# SNOW-1887766 Implement logic to propagate trace context, if it exists, to GS