--- a/connection_test.go
+++ b/connection_test.go
@@ -15,6 +15,9 @@ import (
 	"sync"
 	"testing"
 	"time"
+
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/sdk/trace"
 )
 
 const (
@@ -269,6 +272,58 @@ func TestExecWithSpecificRequestID(t *testing.T) {
 		false /* describeOnly */, nil); err != nil {
 		t.Fatalf("err: %v", err)
 	}
+}
+
+func TestExecContextPropagationIntegrationTest(t *testing.T) {
+	originalTracerProvider := otel.GetTracerProvider()
+	t.Cleanup(func() {
+		otel.SetTracerProvider(originalTracerProvider)
+	})
+
+	tp := trace.NewTracerProvider()
+	otel.SetTracerProvider(tp)
+
+	tracer := otel.Tracer("TestExecContextPropagationTracer")
+
+	ctx, span := tracer.Start(context.Background(), "test-span")
+	defer span.End()
+
+	traceID := span.SpanContext().TraceID().String()
+	spanID := span.SpanContext().SpanID().String()
+
+	// expected header values
+	expectedTraceparent := fmt.Sprintf("00-%s-%s-01", traceID, spanID)
+
+	postQueryMock := func(_ context.Context, _ *snowflakeRestful,
+		_ *url.Values, headers map[string]string, _ []byte, _ time.Duration,
+		_ UUID, _ *Config) (*execResponse, error) {
+
+		// ensure the traceID and spanID from the ctx passed in has been injected into the headers
+		// in W3 Trace Context format
+		assertEqualE(t, headers["traceparent"], expectedTraceparent)
+
+		dd := &execResponseData{}
+		return &execResponse{
+			Data:    *dd,
+			Message: "",
+			Code:    "0",
+			Success: true,
+		}, nil
+	}
+
+	sr := &snowflakeRestful{
+		FuncPostQuery: postQueryMock,
+	}
+
+	sc := &snowflakeConn{
+		cfg:               &Config{Params: map[string]*string{}},
+		rest:              sr,
+		queryContextCache: (&queryContextCache{}).init(),
+	}
+
+	_, err1 := sc.exec(ctx, "", false /* noResult */, false, /* isInternal */
+		false /* describeOnly */, nil)
+	assertNilF(t, err1)
 }
 
 // TestServiceName tests two things: