@sfc-gh-snowflakedb-snyk-sa 2025-04-28
![](https://res.cloudinary.com/snyk/image/upload/r-d/scm-platform/snyk-pull-requests/pr-banner-default.svg)

### :tada: **Snyk checks have passed. No issues have been found so far.**

:white_check_mark: **security/snyk** check is complete. No issues have been found. [(View Details)](https://app.snyk.io/org/snowflakedb-sca-scanning-public-repo/pr-checks/c2a87d13-e601-4fa9-9dc4-094ed4942321?refs=2e64efa12515f83fbb33551a66400f84946225b1&source=prChecksComment)

:white_check_mark: **license/snyk** check is complete. No issues have been found. [(View Details)](https://app.snyk.io/org/snowflakedb-sca-scanning-public-repo/pr-checks/c2a87d13-e601-4fa9-9dc4-094ed4942321/license?refs=2e64efa12515f83fbb33551a66400f84946225b1&source=prChecksComment)